Got formatted alog implementation working Added more ubsan stubs Update TODO.md

+++ alonira/alocom/alomemory.c
@@ -3,12 +3,12 @@
 
 #include "include/alomemory.h"
 
-alo_error_t alo_memory_set(void* const restrict address, const unsigned long length, const unsigned char value) {
+alo_error_t alo_memory_set(void* const restrict address, const size_t length, const unsigned char value) {
 	ALO_FRAME_BEGIN(alo_memory_set);
 
 	ALO_NULL_CHECK(address);
 
-	for(unsigned char* d = address; (unsigned long) (d - (unsigned char*) address) < length; ++d) {
+	for(unsigned char* d = address; (size_t) (d - (unsigned char*) address) < length; ++d) {
 		*d = value;
 	}
