+++ alonira/alocom/include/alostring.h
@@ -6,16 +6,16 @@
 
 #include "alocom.h"
 
-#define ALO_STRING_NO_BOUND ((unsigned long) -1)
+#define ALO_STRING_NO_BOUND ((size_t) -1)
 
-#define ALO_STRING_FOREACH(member, members_length, container) for(register __typeof__((container)[0])* member = &(container)[0]; (unsigned long) ((member) - (container)) < (members_length); ++(member))
+#define ALO_STRING_FOREACH(member, members_length, container) for(register __typeof__((container)[0])* member = &(container)[0]; (size_t) ((member) - (container)) < (members_length); ++(member))
 
-ALO_ERRORABLE alo_string_compare(const char* const restrict a, const unsigned long a_bound, const char* const restrict b, const unsigned long b_bound, const unsigned long limit, bool* const restrict out_equal);
-ALO_ERRORABLE alo_string_copy(char* const restrict destination, const unsigned long destination_bound, const char* const restrict source, const unsigned long source_bound, const unsigned long limit);
-ALO_ERRORABLE alo_string_append(char* const restrict destination, const unsigned long destination_bound, const char* const restrict source, const unsigned long source_bound, const unsigned long limit);
-ALO_ERRORABLE alo_string_length(const char* const restrict string, const unsigned long string_bound, const unsigned long limit, unsigned long* const restrict out_length);
-ALO_ERRORABLE alo_string_character_first(const char* const restrict string, const unsigned long string_bound, const char character, const unsigned long limit, const char* restrict* const restrict out_found);
-ALO_ERRORABLE alo_string_character_last(const char* const restrict string, const unsigned long string_bound, const char character, const unsigned long limit, const char* restrict* const restrict out_found);
-ALO_ERRORABLE alo_string_number(const char* const restrict string, const unsigned long string_bound, const unsigned long limit, unsigned long* const restrict out_number);
+ALO_ERRORABLE alo_string_compare(const char* const restrict a, const size_t a_bound, const char* const restrict b, const size_t b_bound, const size_t limit, bool* const restrict out_equal);
+ALO_ERRORABLE alo_string_copy(char* const restrict destination, const size_t destination_bound, const char* const restrict source, const size_t source_bound, const size_t limit);
+ALO_ERRORABLE alo_string_append(char* const restrict destination, const size_t destination_bound, const char* const restrict source, const size_t source_bound, const size_t limit);
+ALO_ERRORABLE alo_string_length(const char* const restrict string, const size_t string_bound, const size_t limit, size_t* const restrict out_length);
+ALO_ERRORABLE alo_string_character_first(const char* const restrict string, const size_t string_bound, const char character, const size_t limit, const char* restrict* const restrict out_found);
+ALO_ERRORABLE alo_string_character_last(const char* const restrict string, const size_t string_bound, const char character, const size_t limit, const char* restrict* const restrict out_found);
+ALO_ERRORABLE alo_string_number(const char* const restrict string, const size_t string_bound, const size_t limit, size_t* const restrict out_number);
 
 #endif
